Fixed the compiler warnings about the unsigned and signed mismatch

+++ Labs/Lab1/lab1.cpp
@@ -60,8 +60,8 @@ int main()
     vector<string> acceptedStringsDFA1;
     vector<string> acceptedStringsDFA2;
 
-    vector<pair<int, int> > validPositionsForDFA1;
-    vector<pair<int, int> > validPositionsForDFA2;
+    vector<pair<int, int>> validPositionsForDFA1;
+    vector<pair<int, int>> validPositionsForDFA2;
 
     cout << "Enter a string: ";
     getline(cin, testString);
@@ -85,7 +85,7 @@ int main()
     int endPosForDFA1;
     int endPosForDFA2;
 
-    for (int i = 0; i < testString.length(); i++)
+    for (size_t i = 0; i < testString.length(); i++)
     {
         char input = testString[i];
 
@@ -148,7 +148,6 @@ int main()
         }
     }
 
-
     if (acceptedStringsDFA1.empty() && acceptedStringsDFA2.empty())
     {
         cout << "\033[31mREJECTED for both languages" << '\n';
@@ -182,7 +181,7 @@ int main()
     if (!acceptedStringsDFA1.empty())
     {
         cout << "\033[32mACCEPTED string(s) for " << language1 << ":" << '\n';
-        for (int i = 0; i < acceptedStringsDFA1.size(); i++)
+        for (size_t i = 0; i < acceptedStringsDFA1.size(); i++)
         {
             cout << acceptedStringsDFA1[i] << ' ';
         }
@@ -196,7 +195,7 @@ int main()
     if (!acceptedStringsDFA2.empty())
     {
         cout << "\033[32mACCEPTED string(s) for " << language2 << ":" << '\n';
-        for (int i = 0; i < acceptedStringsDFA2.size(); i++)
+        for (size_t i = 0; i < acceptedStringsDFA2.size(); i++)
         {
             cout << acceptedStringsDFA2[i] << ' ';
         }
